+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.1
+- Revised source book addition to `PsiPri: "Korranberg Chronicle: Psion's Primer"`
+- Resolved deprecation warning & updated minimum to dnd5e 4.3
+
 ## 2.0.0
 
 - Major Rewrite for dnd5e system 4.0.0